Separated wfs transaction
