+++ src/editor/transactionhandler.js
@@ -4,103 +4,15 @@
  * ======================================================================== */
 "use strict";
 
-var ol = require('openlayers');
-var $ = require('jquery');
 var viewer = require('../viewer');
-var editsStore = require('./editsstore')();
-var dispatcher = require('./editDispatcher');
+var wfsTransaction = require('./wfstransaction');
+var transactions = {
+  WFS: wfsTransaction
+};
 
-var format = undefined;
-var serializer = undefined;
-
-module.exports = function(options) {
-  format = new ol.format.WFS();
-  serializer = new XMLSerializer();
-  return {
-    wfsTransaction: wfsTransaction
+module.exports = function(transaction, layerName) {
+  var type = viewer.getLayer(layerName).get('type');
+  if (transactions.hasOwnProperty(type)) {
+    transactions[type](transaction, layerName);
   }
-}
-
-function writeWfsTransaction(transObj, options) {
-  var node = format.writeTransaction(transObj.insert, transObj.update, transObj.delete, options);
-  return node;
-}
-
-function wfsTransaction(transObj, layerName) {
-  var srsName = viewer.getProjectionCode();
-  var layer = viewer.getLayer(layerName);
-  var source = viewer.getMapSource()[layer.get('sourceName')];
-  var options = {
-    gmlOptions: {
-      srsName: srsName
-    },
-    featureNS: source.workspace,
-    featurePrefix: source.prefix || source.workspace,
-    featureType: layerName
-  };
-  var node = writeWfsTransaction(transObj, options);
-  $.ajax({
-    type: "POST",
-    url: source.url,
-    data: serializer.serializeToString(node),
-    contentType: 'text/xml',
-    success: success,
-    error: error,
-    context: this
-  });
-
-  function success(data) {
-    var result = readResponse(data);
-    if (result) {
-      if (result.transactionSummary.totalUpdated > 0) {
-        dispatcher.emitChangeFeature({
-          feature: transObj.update,
-          layerName: layerName,
-          status: 'finished',
-          action: 'update'
-        });
-      }
-      if (result.transactionSummary.totalDeleted > 0) {
-        dispatcher.emitChangeFeature({
-          feature: transObj.delete,
-          layerName: layerName,
-          status: 'finished',
-          action: 'delete'
-        });
-      }
-      if (result.transactionSummary.totalInserted > 0) {
-        var feature = transObj.insert;
-        dispatcher.emitChangeFeature({
-          feature: transObj.insert,
-          layerName: layerName,
-          status: 'finished',
-          action: 'insert'
-        });
-        var insertIds = result.insertIds;
-        insertIds.forEach(function(insertId, index) {
-          feature[index].setId(insertId);
-        });
-      }
-    } else {
-      error();
-    }
-  }
-
-  function error(e) {
-    var errorMsg = e ? (e.status + ' ' + e.statusText) : "";
-    alert('Det inträffade ett fel när ändringarna skulle sparas till servern...<br><br>' +
-      errorMsg);
-  }
-}
-
-function readResponse(data) {
-  var result;
-  if (window.Document && data instanceof Document && data.documentElement &&
-    data.documentElement.localName === 'ExceptionReport') {
-    alert(data.getElementsByTagNameNS(
-      'http://www.opengis.net/ows', 'ExceptionText').item(0).textContent);
-  } else {
-    result = format.readTransactionResponse(data);
-  }
-  return result;
 }
